pyproject.toml
[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[project]
name = "zotero-arxiv-daily"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitignore-parser>=0.1.11",
    "tiktoken>=0.8.0",
    "feedparser>=6.0.11",
    "hydra-core>=1.3.2",
    "socksio>=1.0.0",
    "loguru>=0.7.3",
    "pyzotero>=1.10.0",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.2.3",
    "torch",
    "openai>=2.24.0",
    "arxiv>=2.4.0",
    "pymupdf4llm>=0.3.4",
    "pymupdf-layout>=1.27.1",
    "dotenv>=0.9.9",
    "peft>=0.18.1",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.pytest.ini_options]
addopts = "-m 'not ci'"
markers = [
    "ci: tests that only run in CI (require external services)",
]
filterwarnings = [
    "ignore::DeprecationWarning:multiprocessing",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=8.4.1",
]
